댓글 작성하기, props fixed
