+++ src/redux/configStore.js
@@ -5,6 +5,7 @@ import { connectRouter } from "connected-react-router";
 
 import user from "./modules/user";
 import chat from "./modules/chat";
+import comment from "./modules/comment";
 
 export const history = createBrowserHistory();
 
@@ -12,6 +13,7 @@ const rootReducer = combineReducers({
   // 모듈 적어주세요.
   user: user,
   chat: chat,
+  comment: comment,
 
   router: connectRouter(history),
 });
