+++ src/redux/modules/comment.js
@@ -8,15 +8,31 @@ const ADD_COMMENT = "ADD_COMMENT";
 const DEL_COMMENT = "DEL_COMMENT";
 
 //Action Creator
+const getComment = createAction(GET_COMMENTS, (boardId, data) => ({ boardId, data }))
 const addComment = createAction(ADD_COMMENT, (boardId, comment) => ({ boardId, comment }))
 const delComment = createAction(DEL_COMMENT, (boardId, commentId) => ({ boardId, commentId }))
 
 //initialState
 const initialState = {
-    comment_list: {},
+    commentList: {},
 }
 
 //MiddleWare
+const getCommentDB = (boardId) => {
+    return async function (dispatch, getState, { history }) {
+        await apis
+            .getComment(boardId)
+            .then((res) => {
+                console.log(res)
+                dispatch(getComment(boardId, res.data))
+            })
+            .catch((err) => {
+                console.log("댓글불러오기 에러", err)
+            })
+    }
+}
+
+
 const addCommentDB = (boardId, comment) => {
     return async function (dispatch, getState, { history }) {
         if (!boardId) {
@@ -57,10 +73,12 @@ const delCommentDB = (boardId, commentId) => {
 export default handleActions(
     {
         [GET_COMMENTS]: (state, action) => produce(state, (draft) => {
-            draft.comment_list[action.payload.boardId] = action.payload.comment_list
+            const boardId = action.payload.boardId;
+            const data = action.payload.data;
+            draft.commentList[boardId] = data;
         }),
         [ADD_COMMENT]: (state, action) => produce(state, (draft) => {
-            draft.comment_list[action.payload.boardId].unshift(action.payload.comment)
+            draft.commentList[action.payload.boardId].unshift(action.payload.comment)
         }),
         [DEL_COMMENT]: (state, action) => produce(state, (draft) => {
             console.log(action)
@@ -73,6 +91,8 @@ export default handleActions(
 
 //Export Action Creator
 const actionCreators = {
+    getComment,
+    getCommentDB,
     addComment,
     addCommentDB,
     delComment,
